--- pkg/web/calendar.go
@@ -1,1 +1,0 @@
-package web
