+++ pkg/web/utils/handler/handler.go
@@ -1,27 +1,18 @@
 package handler
 
 import (
-	"os"
-
-	"github.com/go-redis/redis/v7"
 	"github.com/gofiber/fiber"
 	"go.uber.org/zap"
-	"go.uber.org/zap/zapcore"
 
-	"bgm38/config"
-	"bgm38/pkg/web/utils"
-	"bgm38/pkg/zapx"
+	"bgm38/pkg/utils/log"
+	logger2 "bgm38/pkg/web/utils/logger"
 )
 
-var _logger *zap.Logger
-
 func LogError(f func(*fiber.Ctx, *zap.Logger) error) func(*fiber.Ctx) {
-	if _logger == nil {
-		_logger = getLogger()
-	}
+	var _logger = log.GetLogger()
 
 	return func(ctx *fiber.Ctx) {
-		s := utils.HeaderFields(ctx)
+		s := logger2.HeaderFields(ctx)
 		// rid := ctx.Fasthttp.Response.Header.Len(fiber.HeaderXRequestID)
 		logger := _logger.With(s)
 		// ctx.Locals()
@@ -31,20 +22,4 @@ func LogError(f func(*fiber.Ctx, *zap.Logger) error) func(*fiber.Ctx) {
 		}
 
 	}
-}
-
-func getLogger() *zap.Logger {
-	ec := zap.NewProductionEncoderConfig()
-	ec.EncodeDuration = zapcore.NanosDurationEncoder
-	ec.EncodeTime = zapcore.ISO8601TimeEncoder
-	enc := zapcore.NewJSONEncoder(ec)
-	return zap.New(zapcore.NewCore(
-		enc,
-		zap.CombineWriteSyncers(zapx.NewRedisSink(&redis.Options{
-			Addr:     config.RedisAddr,
-			Password: config.RedisPassword,
-			PoolSize: 3,
-		}, "bgm38 log v2"), zapcore.AddSync(os.Stdout)),
-		zap.InfoLevel,
-	))
 }
